fixes crashlitycs

+++ eMobilePOSAndroid/src/main/java/com/android/emobilepos/models/orders/Order.java
@@ -401,8 +401,8 @@ public class Order implements Cloneable, Serializable {
         } else {
             int count = 0;
             for (OrderProduct orderProduct : getOrderProducts()) {
-                double consigmentQty = TextUtils.isEmpty(orderProduct.getConsignment_qty()) ? 0 : Integer.parseInt(orderProduct.getConsignment_qty());
-                double qty = Integer.parseInt(orderProduct.getOrdprod_qty());
+                double consigmentQty = TextUtils.isEmpty(orderProduct.getConsignment_qty()) ? 0 : Double.parseDouble(orderProduct.getConsignment_qty());
+                double qty = Double.parseDouble(orderProduct.getOrdprod_qty());
                 if (consigmentQty == 0 && qty > 0) {
                     count++;
                 } else if (consigmentQty > 0 && consigmentQty > qty) {
